Add a build-all.sh script to help building the whole jcae bundle. It's main feature is to create required symlinks. Replace log4j-1.2.8.jar by log4j.jar (see jar META-INF for version).
git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@1388 9c4724a3-da1e-0410-bc68-88f18da35cd4

+++ netbeans-suite/make-dist-Linux.sh
@@ -1,5 +1,6 @@
 #! /bin/sh -x
 
+CASROOT=/home/jerome/OpenCASCADE6.2.0/ros
 JAVA_NAME=jre-6
 VERSION=jcae-@JCAE_VERSION@
 rm -rf $VERSION
@@ -21,17 +22,32 @@ EOF
 
 chmod +x jcae.sh
 
-ln -s ../jre-6-Linux/ $JAVA_NAME
+test -z "$JRE_HOME" && JRE_HOME=../jre-6-Linux
+test ! -d "$JRE_HOME" && JRE_HOME=$(readlink -f $(dirname $(readlink -f $(which java)))/..)	
+CASCADE_LIB=$CASROOT/Linux-jcae/lib
+test ! -d "$CASCADE_LIB" && CASCADE_LIB=$CASROOT/Linux/lib
+
+if [ ! -d "$CASCADE_LIB" ]; then
+	echo "Invalid OpenCASCADE directory. Cannot find $CASCADE_LIB"
+	exit 0
+fi
+
+if [ ! -d "$JRE_HOME" ]; then
+	echo "Invalid JRE directory. Cannot find $JRE_HOME"
+	exit 0
+fi
+
+ln -s $JRE_HOME $JAVA_NAME
 rm -rf $JAVA_NAME/plugin $JAVA_NAME/javaw $JAVA_NAME/lib/i386/client/classes.jsa
 mkdir -p OpenCASCADE6.2.0/ros/Linux/
 ln -s $CASROOT/../LICENSE OpenCASCADE6.2.0/LICENSE
-ln -s $CASROOT/Linux-jcae/lib OpenCASCADE6.2.0/ros/Linux/lib
-unzip ../dist/jcae.zip
+ln -s $CASCADE_LIB OpenCASCADE6.2.0/ros/Linux/lib
+unzip -q ../dist/jcae.zip
 mv jcae tmp
 mv tmp/* .
 rmdir tmp
 
 cd ..
 rm $VERSION-Linux.tar.bz2
-tar cvfj $VERSION-Linux.tar.bz2 --totals --owner 0 --group 0 -h $VERSION
+tar cfj $VERSION-Linux.tar.bz2 --totals --owner 0 --group 0 -h $VERSION
